Thinkr-app/backend/requirements.txt
fastapi==0.110.0
sqlmodel==0.0.16
uvicorn[standard]==0.28.0
cuid2==2.0.0
pyjwt==2.8.0
argon2-cffi==21.2.0